Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
openpyxl==3.1.5
groq==0.11.0
python-multipart==0.0.12
aiofiles==24.1.0
requests
beautifulsoup4
openai
python-dotenv
httpx==0.27.2
